Move listener methods around
and remove start method.
